Add KEYSTONE_OPERATOR_TEMPLATES
This moves the config files into templates and adds logic to
template_util so that we can load them at dev time
and in production when they are in a container.

When containerized a KEYSTONE_OPERATOR_TEMPLATES variable is
set which points to the deployed /usr/share/keystone-operator/templates
location within the container.