Use default from route for initial endpoint
After the keystone route is created and the host value generated
automatically, use that as the value of the initial keystone endpoint.
This way, the user does not need to specify the endpoint value.

The route needs to be created before the bootstrap job, so the order of
those tasks are switched in the controller.

Signed-off-by: James Slagle <jslagle@redhat.com>