Use MariaDBSchema CRD to create DB schema
This updates the keystone controller so that it rely's
on the mariadb-operator's MariaDBSchema to create
databases for us.

Important to note that the implementation uses the
unstructured k8s client so that we don't explicitly
depend on the mariadb-operator code struct's (and associated
golang lib versions).

The old code is removed from the dbsync job and we can
also remove the associated parameters.