Change order to bootstrap service before creating deployment
The operator is running the keystone bootstrap job after the
deployment is up. Because there is also no readiness probe,
this means it's not possible for a client to know if it's supposed
to be working, yet.

This changes the order to expose the service (get the endpoint
URLs) and rund the bootstrap job before the deployment of the
keystone service gets created.