Add IsReady method and PasswordSelector
This allows clients to check if the keystoneAPI is up and running
to serve requests. Right now it returns trues if one replica of
the keystoneapi deployement is up. Could be improved later if needed.

Also adds a PasswordSelector which allows to customize which entry
of a secret holds the DB and admin password. Defaults to
KeystoneDatabasePassword and AdminPassord.

Note: KeystoneDatabasePassword is not yet used as the mariadb-operator
does not support passing specific service DB password.