Split KeystoneService into Service and Endpoint
Some operators require to manage creating service, service user
and assign the service user to the required role in a different
component then register the service endpoints. Also there is the
need to be able to create specific service endpoints via different
components.
Therefore this change intruces a new CRD, KeystoneEndpoint, and
moves the endpoint handling from the current KeystoneService CRD.

The KeystoneEndpoint controller gets the service ID to register
the endpoints by requesting the corresponding KeystoneService
object for the given ServiceName.

The new KeystoneEndpoint CRD allows, as it is today, to register
multiple endpoints in keystone via the endpoint map in the spec,
but allows to split into multiple KeystoneEndpoint CRs.