Enhance endpoint cleanup if Spec.Endpoints changed
If the client adds an endpoint and removes another in the same
update call then the length of the EndpointIDs will be the same
as the length of the Spec.Endpoint but the content will differ.

This situation was only covered via multiple reconcilation runs.

This change allows to detect such a situation by validate
Status.EndpoitIDs against Spec.Endpoints and delete an
endpoint if it does no longer exist in Spec.Endpoints but has
a reference in Status.Endpoints.