Check number of pods from deployment
Instead of using bash to get the number of keystone pods (because we
can't match the name exactly since they have a random string
appended), query the deployment and read the number of replicas there.

This commit also adds a test step that checks what happens when the
keystone service is scaled down to zero.