Remove crc specific parameters from kuttl tests
Remove some fields from the asserts (host in routes and apiEndpoints in
CR spec) that had crc specific values since those make the tests fail if
run in a cluster other than crc. It should be easy to check with
a regex, but kuttl does not support it. If they're really important to
check, can introduce a TestAssert for them and check the value with some
script.