Support adopting KeystoneService from existing DB
During DB adoption we are copying over the keystone DB schema from an
existing DB server. If that schema already contains services then the
keystone_service controller needs to make sure that the ServiceID of a
given service is propagated from the DB to the
KeystoneService.Status.ServiceID field when the KeystoneService CR is
created. Without that the KeystoneService won't get into Ready state.