Rename apiEndpoint key for consistency
The apiEndpoint field in status presents multiple endpoints (internal
and public). This renames the key to the plural, to make the key name
consistent with what is presented in that field.

Note that the old key is still kept because the other operators
require it until keystone-operator in their dependencies is bumped.
It will be remove once version bump is done in all operators.