Name resources according to instance.Name
... instead of using static names. This allows us to make naming more
consistent across operators. Although we don't support deploying
multiple control planes in a single name space, we have to deploy
multiple instances of some CR(eg GlanceAPI) in the single namespace to
deploy a single control plane, so instance.Name is the preferred method
to name resources. This does not care about the naming conflicts
between CRs in case different CRs are deployed with the same name,
because adding prefix causes redundant names. (for example if a user
deploys a KeystoneAPI instance with name keystone, we prefer creating
a deployment with name keystone instead of keystone-keystone).

Note that this change does not change name of client config because
that requires update in OpenStackClient controller.