Fix reconcileDelete during namespace delete
`oc delete namespace openstack` hangs forever because KeystoneAPI is
reconciled in the state when the namespace already being terminated but
the KeystoneAPI CR has no delete timestamp set yet. In this state
KeystoneAPI's reconciler tries to manipulate things in the namespace
(service account, TransportURL, etc) and that is rejected by k8s. This
cause KeystoneAPI to become Ready=False. This then prevent
KeystoneEndpoint and KeystoneService deletion to proceed as they are
waiting for KeystoneAPI to become Ready=true.

This patch implements one possible solution. It assumes if the
KeystoneAPI is being deleted then there is no point removing the
endpoints and users from the keystone DB as the DB will be deleted too.
Therefore the KeystoneEndpoint and KeystoneService simply removes the
finalizers if they detect that KeystoneAPI deleted timestamp is set.

This partially resolves the namespace deletion hangs. (There is another
issue with DNSMasq CR that still blocks the namespace deletion)

Fixes: OSP-27793