Use system scoped admin service client per default
To customize roles for a user, like assigining "SwiftSystemReader"
role to the swift user in the "service" project within the telemetry
operator a system scoped client is required. Since the client is
used for deployment its probably ok to always use a system scoped
client.

Also adds GetScopedAdminServiceClient() if an admin cliend with
a different scope is required.