Restore LastTransitionTime after setting Ready Condition
instance.Status.Conditions.MarkTrue() always sets the current
time as the LastTransitionTime and this results in infinite
loop as the LastTransitionTime changes after every reconcile.

I think this issue is there in many other operator controllers,
but we don't see the issue as the next reconcile is probably too
quick to not change the LastTransitionTime(precision for which is
in seconds). The issue showed up in dataplane controllers as we
reconcile a number of services in a loop for every reconcile of
nodeset.

Jira: OSPRH-8811

Signed-off-by: rabi <ramishra@redhat.com>