Update LastAppliedTopology to TopologyRef interface
While TopologyRef is able to reference Name and Namespace, we currently
only save TopologyRef.Name to .Status.LastAppliedTopology. This patch
aligns both interfaces to use the same type in the same form. In
addition, some logic is moved to infra-operator because is common to all
operators, and this patch also adopts this new form. It also updates the
SpecCore webhook interface to accept Namespace as parameter. It ensures
that we properly validate the referenced topology from
openstack-operator, rejecting requests with a different namespace.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>