Add support for multi-realm federation config
It would be great to use something like extraVols to configure
multi-realm federation.  Unfortunately, the config files for
multi-realm - which need to be stored in a secret because they
contain some sensitive data - have filenames that have URL
encodings, making them invalid as keys in a k8s secret.

So, we have to resort to some trickery to get the right files
mounted.  The solution is as follows:
1. User provides a secret which contains the contents
   {key, json map containing filenames and contents}
2. We parse the contents of that secret and create a new secret
   which has valid keys (1,2,3 .. etc) and the relevant contents
   based on an ordered list of the filenames.
3. We create the mounts and mountpoints as appropriate.

Signed-off-by: Veronika Fisarova <vfisarov@redhat.com>