Add Region parameter in keystoneAPI status
There are services like Glance that need to configure parameters that
are based on the Keystone CR we build. oslo_limits is an example of
that, and we need to access the Region parameter that is defined in
keystone CR spec.
It might be possible to rely on gophercloud and perform os.GetRegion(),
but it seems too convoluted given that we regularly consume GetKeystoneAPI
from the service controllers and we can access this parameter without
building an openstack client (that is based - in any case - on
.Spec.Region defined in keystoneAPI CR).
This patch mirrors the Region parameter in the .Status (according to
the ObservedGeneration pattern), so we can access this field from a
Service operator that needs it via a GetRegion function exposed by
keystoneAPI.

Related: https://issues.redhat.com/browse/OSPRH-18291

Signed-off-by: Francesco Pantano <fpantano@redhat.com>